+++ ovxlib/src/kernel/evis/l2normalizescale_evis.c
@@ -139,11 +139,11 @@ DEF_KERNEL_INITIALIZER(_l2normalizescale_initializer)
         input_fl   = input_attr->dfp.fl;
         if (input_fl >= 0)
         {
-            inputScale = 1.0f / (float) (1 << input_fl);
+            inputScale = 1.0f / (float) ((int64_t)1 << input_fl);
         }
         else
         {
-            inputScale = (float) (1 << -input_fl);
+            inputScale = (float) ((int64_t)1 << -input_fl);
         }
     }
     else if ( VSI_NN_KERNEL_QUANT_ASYMM == input_attr->quant )
@@ -157,11 +157,11 @@ DEF_KERNEL_INITIALIZER(_l2normalizescale_initializer)
         output_fl   = output_attr->dfp.fl;
         if (output_fl >= 0)
         {
-            outputScale = (float) (1 << output_fl);
+            outputScale = (float) ((int64_t)1 << output_fl);
         }
         else
         {
-            outputScale = 1.0f / (float) (1 << -output_fl);
+            outputScale = 1.0f / (float) ((int64_t)1 << -output_fl);
         }
     }
     else if ( VSI_NN_KERNEL_QUANT_ASYMM == output_attr->quant )
